+++ test/deribit/client_test.rb
@@ -5,23 +5,58 @@ class Deribit::ClientTest < Minitest::Test
     @client = Deribit::Client.new testnet: true
   end
 
-  def test_time
+  def test_time_http
     time = @client.time
     assert time.positive?
   end
 
-  def test_ttt
-    test = @client.test
-    # assert test.success
-    # assert test.testnet
+  def test_time_websocket
+    @client.time do |time|
+      assert time.positive?
+      @client.websocket.stop
+    end
   end
 
-  def test_ping
+  def test_enable_heartbeat_websocket
+    @client.enable_heartbeat do |success|
+      assert_equal 'ok', success
+      @client.websocket.stop
+    end
+  end
+
+  # def test_cancel_heartbeat_websocket
+  #   @client.cancel_heartbeat do |success|
+  #     assert_equal 'ok', success
+  #     @client.websocket.stop
+  #   end
+  # end
+
+  def test_test_http
+    test = @client.test
+    assert test.success
+    assert test.testnet
+  end
+
+  def test_test_websocket
+    @client.test do |success|
+      assert_includes success, 'test'
+      @client.websocket.stop
+    end
+  end
+
+  def test_ping_http
     pong = @client.ping
     assert_equal 'pong', pong
   end
 
-  def test_instruments
+  def test_ping_websocket
+    @client.ping do |pong|
+      assert_equal 'pong', pong
+      @client.websocket.stop
+    end
+  end
+
+  def test_instruments_http
     instruments = @client.instruments
     assert instruments.size.positive?
     instrument = instruments.first
@@ -29,7 +64,14 @@ class Deribit::ClientTest < Minitest::Test
     assert_includes instrument.instrumentName, 'BTC'
   end
 
-  def test_currencies
+  def test_instruments_websocket
+    @client.instruments do |instrument|
+      assert_equal 'USD', instrument.currency
+      @client.websocket.stop
+    end
+  end
+
+  def test_currencies_http
     currencies = @client.currencies
     assert currencies.size.positive?
     currency = currencies.first
@@ -37,20 +79,44 @@ class Deribit::ClientTest < Minitest::Test
     assert currency.minConfirmation >= 1
   end
 
-  def test_index
-    index = @client.index
-    assert index.btc.positive?
+  def test_currencies_websocket
+    @client.currencies do |currency|
+      assert currency.minConfirmation.positive?
+      assert currency.txFee.positive?
+      @client.websocket.stop
+    end
+  end
+
+  def test_index_http
+    index = @client.index 'ETH'
+    assert index.eth.positive?
     assert index.edp.positive?
   end
 
-  def test_orderbook
+  def test_index_websocket
+    @client.index 'BTC' do |index|
+      assert index.btc.positive?
+      @client.websocket.stop
+    end
+  end
+
+  def test_orderbook_http
     orderbook = @client.orderbook 'BTC-PERPETUAL', depth: 3
     assert_equal 'BTC-PERPETUAL', orderbook.instrument
     assert_equal 3, orderbook.bids.size
     assert_equal 3, orderbook.asks.size
   end
 
-  def test_trades
+  def test_orderbook_websocket
+    @client.orderbook 'ETH-PERPETUAL', depth: 3 do |orderbook|
+      assert_equal 'ETH-PERPETUAL', orderbook.instrument
+      assert_equal 3, orderbook.bids.size
+      assert_equal 3, orderbook.asks.size
+      @client.websocket.stop
+    end
+  end
+
+  def test_trades_http
     trades = @client.trades 'options', count: 3
     assert_equal 3, trades.size
     trade = trades.first
@@ -59,7 +125,15 @@ class Deribit::ClientTest < Minitest::Test
     assert trade.iv.positive?
   end
 
-  def test_summary
+  def test_trades_websocket
+    @client.trades do |trade|
+      assert trade.price.positive?
+      assert trade.quantity.positive?
+      @client.websocket.stop
+    end
+  end
+
+  def test_summary_http
     summaries = @client.summary :futures
     summary = summaries.first
     assert summary.openInterest.positive?
@@ -67,7 +141,14 @@ class Deribit::ClientTest < Minitest::Test
     assert summary.volumeBtc.positive?
   end
 
-  def test_stats
+  def test_summary_websocket
+    @client.summary do |summary|
+      assert summary.instrumentName
+      @client.websocket.stop
+    end
+  end
+
+  def test_stats_http
     stats = @client.stats
     assert_kind_of Hash, stats
     btcusd = stats.btc_usd
@@ -76,23 +157,48 @@ class Deribit::ClientTest < Minitest::Test
     assert btcusd.putsVolume.positive?
   end
 
-  def  test_announcements
+  def test_stats_websocket
+    @client.stats do |stats|
+      btcusd = stats.btc_usd
+      assert btcusd.futuresVolume.positive?
+      assert btcusd.callsVolume.positive?
+      assert btcusd.putsVolume.positive?
+      @client.websocket.stop
+    end
+  end
+
+  def  test_announcements_http
     announcements = @client.announcements
     assert announcements.size.positive?
     assert !announcements.first.title.nil?
   end
 
-  def test_settlements
-    skip 'it fails on testnet, method was removed?'
-    settlements = @client.settlements
-    assert settlements.size.positive?
-    assert settlements.first.indexPrice.positive?
+  def test_announcements_websocket
+    @client.announcements do |announcement|
+      assert !announcement.title.nil?
+      @client.websocket.stop
+    end
   end
 
+  def test_settlements_http
+    settlements = @client.settlements
+    assert !settlements.continuation.nil?
+    assert settlements.settlements.size.positive?
+    assert settlements.settlements.first.indexPrice.positive?
+  end
+
+  def test_settlements_websocket
+    @client.settlements do |settlements|
+      assert !settlements.continuation.nil?
+      assert settlements.settlements.size.positive?
+      assert settlements.settlements.first.indexPrice.positive?
+      @client.websocket.stop
+    end
+  end
 
   class PrivateTests < Minitest::Test
     def  setup
-      @client = Deribit::Client.new key: ENV['API_KEY'], secret: ENV['API_SECRET'], testnet: true
+      @client = Deribit::Client.new key: ENV['API_KEY'], secret: ENV['API_SECRET'], testnet: true, debug: false
     end
 
     def test_account
@@ -159,7 +265,8 @@ class Deribit::ClientTest < Minitest::Test
     end
 
     def test_trades_history
-      trades = @client.trades_history
+      skip 'it fails on testnet even if there are history trades '
+      trades = @client.trades_history :all, startTimestamp: Time.new(2019, 3, 1).to_i
       assert trades.size.positive?
     end