+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- public Websocket API
+
 ## [0.0.3] - 2019-03-06
 ### Added
 - private HTTP API endpoints